backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
httpx>=0.25.0
openai>=1.0.0
python-multipart>=0.0.6
beautifulsoup4>=4.12.2
chro madb>=0.4.22
asyncio>=3.4.3
lxml>=4.9.3
redis>=5.0.1
APScheduler>=3.10.4
requests>=2.31.0
langchain>=0.2.1
langchain-openai>=0.2.1
langchain-community>=0.2.1
langchain-core>=0.2.1
azure-cosmos==4.5.0
aiohttp==3.8.5